Draft  Complete
GUI is designed. Event handlers need the proper code.

Program is able to read the FSA textfile, which contains the rules for the program to follow.

Program is able to verify a pattern is in the language set upon by the textfile.

Need to clean up on of the files and turn it into a class.

Comments are needed.